requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
psycopg2-binary>=2.9
clickhouse-connect>=0.7
redis>=5.0
django-prometheus>=2.3
gunicorn>=22.0
requests>=2.31
pytest>=8.0
pytest-django>=4.8
# apache-airflow is pre-installed in Dockerfile.airflow image
